+++ packages/eyes-sdk-core/test/it/new/close.spec.js
@@ -1,7 +1,7 @@
 const assert = require('assert')
 const assertRejects = require('assert-rejects')
 const {startFakeEyesServer} = require('@applitools/sdk-fake-eyes-server')
-const {MockDriver, fake} = require('@applitools/driver')
+const {MockDriver, spec} = require('@applitools/driver/fake')
 const {generateScreenshot} = require('../../utils/FakeScreenshot')
 const makeSDK = require('../../../lib/new/sdk')
 
@@ -18,7 +18,7 @@ describe('close', async () => {
       {selector: 'element3', rect: {x: 30, y: 31, width: 301, height: 302}},
       {selector: 'element4', rect: {x: 40, y: 41, width: 401, height: 402}},
     ])
-    const core = new makeSDK({spec: fake.spec})
+    const core = new makeSDK({spec})
     server = await startFakeEyesServer({logger: {log: () => {}}, matchMode: 'never'})
     serverUrl = `http://localhost:${server.port}`
     manager = await core.makeManager()