+++ packages/eyes-sdk-core/test/unit/EyesClassic.spec.js
@@ -3,7 +3,7 @@
 const assert = require('assert')
 const assertRejects = require('assert-rejects')
 const {startFakeEyesServer} = require('@applitools/sdk-fake-eyes-server')
-const {MockDriver} = require('@applitools/driver')
+const {MockDriver} = require('@applitools/driver/fake')
 const {EyesClassic} = require('../utils/FakeSDK')
 const {generateScreenshot} = require('../utils/FakeScreenshot')