+++ packages/eyes-sdk-core/test/utils/FakeSDK.js
@@ -1,10 +1,10 @@
-const {fake} = require('@applitools/driver')
+const {spec} = require('@applitools/driver/fake')
 const VisualGridClient = require('@applitools/visual-grid-client')
 const {EyesSDK} = require('../../index')
 
 module.exports = EyesSDK({
   name: 'eyes.fake',
   version: '0.2.0',
-  spec: fake.spec,
+  spec,
   VisualGridClient,
 })